genesAndProteins (postLoadGenome)

+++ Main/lib/perl/WorkflowSteps/CopyIndexedFileToWebServiceDir.pm
@@ -44,11 +44,17 @@ sub run {
 
   if($undo) {
     $self->runCmd(0, "rm -f $copyToDir/$gzFile");
-    $self->runCmd(0, "rm -f $copyToDir/$tbiFile");
+
+    if($indexSuffix) {
+      $self->runCmd(0, "rm -f $copyToDir/$tbiFile");
+    }
+
   } else{
     $self->runCmd($test, "mkdir -p $copyToDir");
     $self->runCmd($test, "cp $workflowDataDir/${copyFromDirName}/${gzFile} $copyToDir");
-    $self->runCmd($test, "cp $workflowDataDir/${copyFromDirName}/${tbiFile} $copyToDir");
+    if($indexSuffix) {
+      $self->runCmd($test, "cp $workflowDataDir/${copyFromDirName}/${tbiFile} $copyToDir");
+    }
   }
 }
